extract modules and adds Storage as the extension cls
