+++ flask_storage/__init__.py
@@ -1,0 +1,2 @@
+from .storage import Storage  # noqa
+from .upload_set import UploadSet  # noqa
